+++ Mortz.Core/Chat/ChatPolicy.cs
@@ -6,7 +6,7 @@ public sealed class ChatPolicy
 {
     private readonly PeerRateLimiter _limiter = new(capacity: 5, tokensPerSecond: 1);
 
-    public bool TryAccept(long peerId, ulong nowMs, string? value,
+    public bool TryAccept(int peerId, ulong nowMs, string? value,
         out string sanitized, out ChatRejectReason reason)
     {
         if (!_limiter.Allow(peerId, nowMs))
@@ -19,8 +19,8 @@ public sealed class ChatPolicy
     }
 
     /// <summary>A roll is a chat line, so it spends the same rate budget.</summary>
-    public bool TryAcceptRoll(long peerId, ulong nowMs) => _limiter.Allow(peerId, nowMs);
+    public bool TryAcceptRoll(int peerId, ulong nowMs) => _limiter.Allow(peerId, nowMs);
 
-    public void Remove(long peerId) => _limiter.Remove(peerId);
+    public void Remove(int peerId) => _limiter.Remove(peerId);
     public void Reset() => _limiter.Reset();
 }
