+++ Mortz.Core/Match/SwapOffer.cs
@@ -1,4 +1,4 @@
 namespace Mortz.Core.Match;
 
 /// <summary>A pending cross-team trade.</summary>
-public readonly record struct SwapOffer(long From, long To);
+public readonly record struct SwapOffer(int From, int To);
