+++ Mortz.Core/Net/Messages/RosterMsg.cs
@@ -5,4 +5,4 @@ namespace Mortz.Core.Net.Messages;
 /// in-game join/leave; the lobby uses LobbyStateMsg instead.</summary>
 [NetMessage(NetChannel.RELIABLE, NetDirection.SERVER_TO_CLIENT)]
 public readonly partial record struct RosterMsg(
-    long[] PeerIds, string[] Names, byte[] Skins, byte[] Teams, byte[] Slots);
+    int[] PeerIds, string[] Names, byte[] Skins, byte[] Teams, byte[] Slots);
