+++ Mortz.Core/Net/Messages/SessionWinsMsg.cs
@@ -3,4 +3,4 @@ namespace Mortz.Core.Net.Messages;
 /// <summary>Parallel arrays: match wins per player for the current server
 /// session. Sent to a joining peer and broadcast whenever a match is won.</summary>
 [NetMessage(NetChannel.RELIABLE, NetDirection.SERVER_TO_CLIENT)]
-public readonly partial record struct SessionWinsMsg(long[] PeerIds, int[] Wins);
+public readonly partial record struct SessionWinsMsg(int[] PeerIds, int[] Wins);
