+++ Mortz.Core/Net/Messages/TeamSwapRequestMsg.cs
@@ -4,4 +4,4 @@ namespace Mortz.Core.Net.Messages;
 /// Repeating it cancels the offer; the target offering back executes the
 /// swap. The server answers with the lobby broadcast either way.</summary>
 [NetMessage(NetChannel.RELIABLE, NetDirection.CLIENT_TO_SERVER)]
-public readonly partial record struct TeamSwapRequestMsg(long TargetPeerId);
+public readonly partial record struct TeamSwapRequestMsg(int TargetPeerId);
