+++ Mortz.Core/Net/NetTransport.cs
@@ -8,11 +8,11 @@ namespace Mortz.Core.Net;
 public static class NetTransport
 {
     /// <summary>Send target: every validated peer.</summary>
-    public const long BROADCAST = 0;
+    public const int BROADCAST = 0;
     /// <summary>Send target: the server (Godot peer id 1).</summary>
-    public const long TO_SERVER = 1;
+    public const int TO_SERVER = 1;
 
-    public delegate void SendDelegate(ushort msgId, byte[] payload, long target, NetChannel channel);
+    public delegate void SendDelegate(ushort msgId, byte[] payload, int target, NetChannel channel);
 
     public static SendDelegate Send { get; set; } =
         static (_, _, _, _) => throw new InvalidOperationException("NetTransport.Send not assigned");
