+++ Mortz.Core/Net/PeerAdmissionState.cs
@@ -4,39 +4,39 @@ namespace Mortz.Core.Net;
 public sealed class PeerAdmissionState
 {
     private readonly ulong _timeoutMs;
-    private readonly Dictionary<long, ulong> _pending = new();
-    private readonly HashSet<long> _validated = new();
+    private readonly Dictionary<int, ulong> _pending = new();
+    private readonly HashSet<int> _validated = new();
 
     public PeerAdmissionState(ulong timeoutMs = NetConfig.HELLO_TIMEOUT_MS) => _timeoutMs = timeoutMs;
 
-    public IReadOnlyCollection<long> ValidatedPeers => _validated;
-    public bool IsValidated(long peerId) => _validated.Contains(peerId);
+    public IReadOnlyCollection<int> ValidatedPeers => _validated;
+    public bool IsValidated(int peerId) => _validated.Contains(peerId);
 
-    public void Connected(long peerId, ulong nowMs)
+    public void Connected(int peerId, ulong nowMs)
     {
         _validated.Remove(peerId);
         _pending[peerId] = SaturatingAdd(nowMs, _timeoutMs);
     }
 
     /// <summary>True only for the first Hello from a connected pending peer.</summary>
-    public bool TryValidate(long peerId)
+    public bool TryValidate(int peerId)
     {
         if (!_pending.Remove(peerId))
             return false;
         return _validated.Add(peerId);
     }
 
-    public long[] Expire(ulong nowMs)
+    public int[] Expire(ulong nowMs)
     {
-        long[] expired = _pending.Where(pair => pair.Value <= nowMs).Select(pair => pair.Key).ToArray();
-        foreach (long peerId in expired)
+        int[] expired = _pending.Where(pair => pair.Value <= nowMs).Select(pair => pair.Key).ToArray();
+        foreach (int peerId in expired)
         {
             _pending.Remove(peerId);
         }
         return expired;
     }
 
-    public bool Remove(long peerId)
+    public bool Remove(int peerId)
     {
         _pending.Remove(peerId);
         return _validated.Remove(peerId);
