+++ Mortz.Core/Net/RosterEntry.cs
@@ -4,7 +4,7 @@ namespace Mortz.Core.Net;
 
 public readonly record struct RosterEntry
 {
-    public RosterEntry(long peerId, string name, byte skin, Team? team, NetSlot slot)
+    public RosterEntry(int peerId, string name, byte skin, Team? team, NetSlot slot)
     {
         if (peerId <= 0)
             throw new ArgumentOutOfRangeException(nameof(peerId));
@@ -15,7 +15,7 @@ public readonly record struct RosterEntry
         Slot = slot;
     }
 
-    public long PeerId { get; }
+    public int PeerId { get; }
     public string Name { get; }
     public byte Skin { get; }
     public Team? Team { get; }
