+++ Mortz.Core/Net/ScoreRow.cs
@@ -1,3 +1,3 @@
 namespace Mortz.Core.Net;
 
-public readonly record struct ScoreRow(long PeerId, int Kills, int Deaths);
+public readonly record struct ScoreRow(int PeerId, int Kills, int Deaths);
