+++ src/Client/Announcements/Combatant.cs
@@ -4,4 +4,4 @@ namespace Mortz.Client.Announcements;
 
 /// <summary>A player as the roster knew them when the event landed. Team is
 /// null when teams are off.</summary>
-public readonly record struct Combatant(long Id, string Name, Team? Team);
+public readonly record struct Combatant(int Id, string Name, Team? Team);
